requirements.txt
requests
rsa
pyyaml
